model-optimizer/requirements_tf.txt
tensorflow>=1.2.0
networkx>=1.11
numpy>=1.12.0
test-generator==0.1.1
defusedxml>=0.5.0